Simulation for Lab 2
main.cpp does everything needed for Lab 2 (I think). There's a base Station class and subclasses for the 3 different kinds of sensors. The biggest issue I can find is that my average delays for the non-persistent case are consistently LOWER than the 1-Persistent case, which strikes me as odd. This might be because signifcantly fewer packets are transmitted, but I'm not sure